Fix re-reading an eventfd
Add a test that fails on the latest version of the code.

Update `read` to match the example code for `AsyncTcpStream` in the
`AsyncFd` documentation.  This fixes the test.

I'm not 100% sure what's wrong with the original code.  I think maybe
after getting EWOULDBLOCK from the `inner.read()` call, we need to loop
back up to call `poll_read_ready` so tokio knows to re-arm us for
receiving read-ready events.